pyproject.toml
[project]
name = "gallery"
version = "0.1.0"
description = ""
authors = [
    {name = "Timothy Wong",email = "timothy.wong@hotmail.co.uk"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow (>=11.3.0,<12.0.0)",
    "fastapi (>=0.112.0,<1.0.0)",
    "uvicorn[standard] (>=0.30.0,<0.31.0)",
    "jinja2 (>=3.1.0,<4.0.0)",
    "python-multipart (>=0.0.9,<0.0.10)",
    "httpx (>=0.24.0,<1.0.0)"
]

[project.scripts]
gallery = "app.main:main"

[tool.poetry]
packages = [{ include = "app" }]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
